Bug(Client): Protect pages from unauthorized access
Move the add sign up a bit for mobile devices
Protect add page, get one page, modify page and profile page
Add home icon to the profile page

[Delivers #159795312]

+++ client/js/profile.js
@@ -1,13 +1,17 @@
-window.addEventListener('load', () => {
-  const signOut = document.getElementById('signOut');
-  const entriesNumber = document.getElementById('entries');
+if (sessionStorage.getItem('token')) {
+  window.addEventListener('load', () => {
+    const signOut = document.getElementById('signOut');
+    const entriesNumber = document.getElementById('entries');
 
-  entriesNumber.innerHTML = `You have a total of ${sessionStorage.getItem('entriesTotal')} entries`;
+    entriesNumber.innerHTML = `You have a total of ${sessionStorage.getItem('entriesTotal')} entries`;
 
-  signOut.addEventListener('click', () => {
-    signOut.value = '';
-    signOut.style.background = 'rgb(20, 43, 68) url(images/Spinner-1s.gif) no-repeat center';
-    sessionStorage.clear();
-    window.location.href = 'index.html';
+    signOut.addEventListener('click', () => {
+      signOut.value = '';
+      signOut.style.background = 'rgb(20, 43, 68) url(images/Spinner-1s.gif) no-repeat center';
+      sessionStorage.clear();
+      window.location.href = 'index.html';
+    });
   });
-});
+} else {
+  window.location.href = 'index.html';
+}
